refactor(calculator): extract PartsSection into its own component
The collapsible Parts / materials card is fully self-contained (only
touches the parts list and a show/hide flag), so it's a clean extraction
that further trims the parent's JSX.
