+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tech-support-site",
-  "version": "1.38.6",
+  "version": "1.38.7",
   "private": true,
   "type": "module",
   "scripts": {
